Rework of the logistic controller and into of a PID controller
Add a simple (manually tunned) PID controller for the fan duty
cycle based on the monitored thermal device.

The logic of all controllers was modified to allow a clean
implementation of the two existing ones and any future controllers.

Introduction of a new arg (-o) that lets users choose a thermal
controller for the fan ('logistic' or 'pid').

Key controller parameters were added to the top of the sript
along with other global variables. This makes it easier to edit
them without going into the specifics of their implementation.

Increased default fan startup time to gather a more reliable metric
of the initial temperature to be used by the PID controller as the
ideal/desired one.